+++ app/controllers/users_controller.rb
@@ -5,23 +5,21 @@ class UsersController < ApplicationController
   before_action :require_login, except: %w[new create]
   before_action :require_logout, only: %w[new create]
 
-  # GET /users
-  # GET /users.json
-  def index
-    @users = User.all
-  end
-
-  # GET /users/1
-  # GET /users/1.json
   def show; end
+
+  def profile
+    @user = logged_user
+    render 'users/show'
+  end
 
   def new
     @user = User.new
     render 'auth/signup'
   end
 
-  # GET /users/1/edit
-  def edit; end
+  def edit
+    redirect_to root_path unless logged_user == @user
+  end
 
   def create
     @user = User.new(user_params)
@@ -34,28 +32,21 @@ class UsersController < ApplicationController
     end
   end
 
-  # PATCH/PUT /users/1
-  # PATCH/PUT /users/1.json
   def update
-    respond_to do |format|
-      if @user.update(user_params)
-        format.html { redirect_to @user, notice: 'User was successfully updated.' }
-        format.json { render :show, status: :ok, location: @user }
-      else
-        format.html { render :edit }
-        format.json { render json: @user.errors, status: :unprocessable_entity }
-      end
+    redirect_back fallback_location: root_path unless logged_user == @user
+
+    if @user.update(user_params)
+      redirect_to root_path
+    else
+      redirect_back fallback_location: edit_user_path
     end
   end
 
-  # DELETE /users/1
-  # DELETE /users/1.json
   def destroy
-    @user.destroy
-    respond_to do |format|
-      format.html { redirect_to users_url, notice: 'User was successfully destroyed.' }
-      format.json { head :no_content }
-    end
+    logged_user.destroy
+    logout
+
+    redirect_to login_path
   end
 
   private