+++ config/routes.rb
@@ -1,10 +1,10 @@
 Rails.application.routes.draw do
-  root 'users#index'
+  root 'users#profile'
   
   get '/signup', to: 'users#new'
   post '/signup', to: 'users#create'
   get '/login', to: 'sessions#new'
   post '/login', to: 'sessions#create'
   get '/logout', to: 'sessions#destroy'
-  resources :users, only: ['show', 'edit', 'update', 'destroy', 'index']
+  resources :users, only: ['show', 'edit', 'update', 'destroy']
 end